move test utils to testutils package
