+++ pkg/diagnostics/component_monitoring_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/dapr/dapr/pkg/diagnostics/utils"
+	"github.com/dapr/dapr/pkg/diagnostics/testutils"
 
 	"github.com/stretchr/testify/assert"
 	"go.opencensus.io/stats/view"
@@ -31,7 +31,7 @@ func TestPubSub(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/pubsub_ingress/count")
 		v := view.Find("component/pubsub_ingress/count")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 	})
 
 	t.Run("record ingress latency", func(t *testing.T) {
@@ -42,7 +42,7 @@ func TestPubSub(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/pubsub_ingress/latencies")
 		v := view.Find("component/pubsub_ingress/latencies")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 
 		assert.Equal(t, float64(1), viewData[0].Data.(*view.DistributionData).Min)
 	})
@@ -55,7 +55,7 @@ func TestPubSub(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/pubsub_egress/latencies")
 		v := view.Find("component/pubsub_egress/latencies")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 
 		assert.Equal(t, float64(1), viewData[0].Data.(*view.DistributionData).Min)
 	})
@@ -70,7 +70,7 @@ func TestBindings(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/input_binding/count")
 		v := view.Find("component/input_binding/count")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 	})
 
 	t.Run("record input binding latency", func(t *testing.T) {
@@ -81,7 +81,7 @@ func TestBindings(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/input_binding/latencies")
 		v := view.Find("component/input_binding/count")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 
 		assert.Equal(t, float64(1), viewData[0].Data.(*view.DistributionData).Min)
 	})
@@ -94,7 +94,7 @@ func TestBindings(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/output_binding/count")
 		v := view.Find("component/input_binding/count")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 	})
 
 	t.Run("record output binding latency", func(t *testing.T) {
@@ -105,7 +105,7 @@ func TestBindings(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/output_binding/latencies")
 		v := view.Find("component/output_binding/latencies")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 
 		assert.Equal(t, float64(1), viewData[0].Data.(*view.DistributionData).Min)
 	})
@@ -120,7 +120,7 @@ func TestState(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/state/count")
 		v := view.Find("component/state/count")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 	})
 
 	t.Run("record state latency", func(t *testing.T) {
@@ -131,7 +131,7 @@ func TestState(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/state/latencies")
 		v := view.Find("component/state/latencies")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 		assert.Equal(t, float64(1), viewData[0].Data.(*view.DistributionData).Min)
 	})
 }
@@ -145,7 +145,7 @@ func TestConfiguration(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/configuration/count")
 		v := view.Find("component/configuration/count")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 	})
 
 	t.Run("record configuration latency", func(t *testing.T) {
@@ -156,7 +156,7 @@ func TestConfiguration(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/configuration/latencies")
 		v := view.Find("component/configuration/latencies")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 
 		assert.Equal(t, float64(1), viewData[0].Data.(*view.DistributionData).Min)
 	})
@@ -171,7 +171,7 @@ func TestSecrets(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/secret/count")
 		v := view.Find("component/secret/count")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 	})
 
 	t.Run("record secret latency", func(t *testing.T) {
@@ -182,7 +182,7 @@ func TestSecrets(t *testing.T) {
 		viewData, _ := view.RetrieveData("component/secret/latencies")
 		v := view.Find("component/secret/latencies")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 
 		assert.Equal(t, float64(1), viewData[0].Data.(*view.DistributionData).Min)
 	})
