+++ pkg/diagnostics/service_monitoring_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/dapr/dapr/pkg/diagnostics/utils"
+	"github.com/dapr/dapr/pkg/diagnostics/testutils"
 
 	"github.com/stretchr/testify/assert"
 	"go.opencensus.io/stats/view"
@@ -26,7 +26,7 @@ func TestServiceInvocation(t *testing.T) {
 		viewData, _ := view.RetrieveData("runtime/service_invocation/req_sent_total")
 		v := view.Find("runtime/service_invocation/req_sent_total")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 	})
 
 	t.Run("record service invoation request received", func(t *testing.T) {
@@ -37,7 +37,7 @@ func TestServiceInvocation(t *testing.T) {
 		viewData, _ := view.RetrieveData("runtime/service_invocation/req_recv_total")
 		v := view.Find("runtime/service_invocation/req_recv_total")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 	})
 
 	t.Run("record service invocation response sent", func(t *testing.T) {
@@ -48,7 +48,7 @@ func TestServiceInvocation(t *testing.T) {
 		viewData, _ := view.RetrieveData("runtime/service_invocation/res_sent_total")
 		v := view.Find("runtime/service_invocation/res_sent_total")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 	})
 
 	t.Run("record service invocation response received", func(t *testing.T) {
@@ -59,12 +59,12 @@ func TestServiceInvocation(t *testing.T) {
 		viewData, _ := view.RetrieveData("runtime/service_invocation/res_recv_total")
 		v := view.Find("runtime/service_invocation/res_recv_total")
 
-		utils.AllTagsPresent(t, v, viewData[0].Tags)
+		testutils.AllTagsPresent(t, v, viewData[0].Tags)
 
 		viewData2, _ := view.RetrieveData("runtime/service_invocation/res_recv_latency_ms")
 		v2 := view.Find("runtime/service_invocation/res_recv_latency_ms")
 
-		utils.AllTagsPresent(t, v2, viewData2[0].Tags)
+		testutils.AllTagsPresent(t, v2, viewData2[0].Tags)
 	})
 }
